requirements.txt
## Code Quality
flake8
black
pytest

## Dev Dependencies
numpy
pandas
tqdm
matplotlib
seaborn
scikit-learn
imgaug
exifread
openpyxl
Pillow
tf-keras-vis
fiftyone

# MLOps
mlflow
sahi
ultralytics


# Deep Learning req
tensorflow
nvidia-cudnn-cu12
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
PyYAML==5.1
scikit-image==0.22.0
